# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API test-backend-developer",
        "description": "Listado de URI's de la API test-backend-developer",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/createVenta": {
            "post": {
                "tags": [
                    "Venta"
                ],
                "summary": "Crear una nueva venta",
                "operationId": "App\\Http\\Controllers\\VentaController::createVenta",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_vendedor": {
                                        "description": "Crear una nueva venta",
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "id_cliente": {
                                        "description": "Crear una nueva venta",
                                        "type": "string",
                                        "example": "2"
                                    },
                                    "id_sucursal": {
                                        "description": "Crear una nueva venta",
                                        "type": "string",
                                        "example": "3"
                                    },
                                    "fecha_venta": {
                                        "description": "Crear una nueva venta",
                                        "type": "string",
                                        "example": "2014-03-10"
                                    },
                                    "monto_total": {
                                        "description": "Crear una nueva venta",
                                        "type": "number",
                                        "example": "123.45"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Venta creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Crear una nueva venta",
                                            "type": "string",
                                            "example": "Venta creada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Crear una nueva venta",
                                            "type": "string",
                                            "example": "Error en los datos enviados"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}